requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
albucore=0.0.12=pypi_0
albumentations=1.4.10=pypi_0
annotated-types=0.7.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.1.0=pypi_0
ca-certificates=2024.3.11=haa95532_0
certifi=2024.6.2=pypi_0
charset-normalizer=3.3.2=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pyhd8ed1ab_0
debugpy=1.6.7=py39hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
docopt=0.6.2=pypi_0
exceptiongroup=1.2.0=py39haa95532_0
executing=2.0.1=pyhd8ed1ab_0
fastjsonschema=2.20.0=pypi_0
filelock=3.13.1=pypi_0
flatbuffers=24.3.25=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
grpcio=1.64.1=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=pypi_0
imageio=2.34.1=pypi_0
importlib-metadata=7.2.0=pyha770c72_0
importlib_metadata=7.2.0=hd8ed1ab_0
intel-openmp=2021.4.0=pypi_0
ipykernel=6.29.4=pyh4bbf305_0
ipython=8.12.3=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
joblib=1.4.2=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=8.6.2=pyhd8ed1ab_0
jupyter_core=5.7.2=py39hcbf5309_0
jupyterlab-pygments=0.3.0=pypi_0
lazy-loader=0.4=pypi_0
libsodium=1.0.18=h8d14728_1
lightning=2.3.0=pypi_0
lightning-utilities=0.11.2=pypi_0
markdown=3.6=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mistune=3.0.2=pypi_0
mkl=2021.4.0=pypi_0
ml-dtypes=0.4.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
numpy=1.26.3=pypi_0
onnx=1.16.1=pypi_0
onnxruntime=1.18.1=pypi_0
onnxscript=0.1.0.dev20240625=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.0.14=h827c3e9_0
packaging=24.1=pyhd8ed1ab_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=24.0=py39haa95532_0
pipreqs=0.5.0=pypi_0
platformdirs=4.2.2=pyhd8ed1ab_0
prompt-toolkit=3.0.47=pyha770c72_0
prompt_toolkit=3.0.47=hd8ed1ab_0
protobuf=4.25.3=pypi_0
psutil=5.9.0=py39h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pydantic=2.7.4=pypi_0
pydantic-core=2.18.4=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyreadline3=3.4.1=pypi_0
python=3.9.19=h1aa4202_1
python-dateutil=2.9.0=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch-lightning=2.3.0=pypi_0
pywin32=305=py39h2bbff1b_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py39hd77b12b_0
referencing=0.35.1=pypi_0
requests=2.32.3=pypi_0
rpds-py=0.18.1=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.1=pypi_0
setuptools=69.5.1=py39haa95532_0
six=1.16.0=pyh6c4a22f_0
soupsieve=2.5=pypi_0
sqlite=3.45.3=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tbb=2021.11.0=pypi_0
tensorboard=2.17.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.6.18=pypi_0
tinycss2=1.3.0=pypi_0
tomli=2.0.1=pypi_0
torch=2.3.1+cu118=pypi_0
torchaudio=2.3.1+cu118=pypi_0
torchmetrics=1.4.0.post0=pypi_0
torchvision=0.18.1+cu118=pypi_0
tornado=6.2=py39hb82d6ee_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
typing-extensions=4.9.0=pypi_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=pypi_0
vc=14.2=h2eaa2aa_1
vs2015_runtime=14.29.30133=h43f2093_3
wcwidth=0.2.13=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py39haa95532_0
yarg=0.1.9=pypi_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=hd77b12b_0
zipp=3.19.2=pyhd8ed1ab_0
